requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
python-telegram-bot==20.8
aiohttp==3.10.5
python-dotenv==1.0.1
requests==2.32.3
pymysql
flask-swagger-ui==5.21.0
